extract reducer into own file
